fixDataType

+++ server/database/migrations/1713870292477_expenses.ts
@@ -6,7 +6,7 @@ export default class extends BaseSchema {
   public async up () {
     this.schema.createTable(this.tableName, (table) => {
       table.integer('id')
-      table.float('balance')
+      table.double('balance')
       table.timestamp('created_at', { useTz: true })
       table.timestamp('updated_at', { useTz: true }).notNullable().defaultTo(this.now())
     })
